Move the profile related code to Features/Profile
