+++ src/API/LeadershipProfileAPI/Features/Profile/Extensions.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Net.Http.Headers;
 
-namespace LeadershipProfileAPI.Controllers
+namespace LeadershipProfileAPI.Features.Profile
 {
     public static class Extensions
     {
